+++ api/index.php
@@ -2,26 +2,9 @@
 
 namespace Bifrost;
 
+require_once __DIR__ . "/Core/Autoload.php";
+
 use Bifrost\Core\Request;
-
-/**
- * Função responsável por importar classes do sistema.
- *
- * @param string $className O nome da classe a ser importada.
- * @return bool Retorna true se a classe foi importada com sucesso, caso contrário retorna false.
- */
-spl_autoload_register(
-    function (string $className): bool {
-        $prefix = 'Bifrost\\';
-        if (strpos($className, $prefix) === 0) {
-            $className = substr($className, strlen($prefix));
-        }
-
-        $file = __DIR__ . DIRECTORY_SEPARATOR . str_replace("\\", DIRECTORY_SEPARATOR, $className) . ".php";
-        require_once $file;
-        return true;
-    }
-);
 
 /**
  * Este arquivo é o ponto de entrada para o servidor.
